requirements.txt
python-telegram-bot>=20.0
requests==2.31.0
gspread==5.8.0
google-auth==2.17.3
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3